fix: address remaining PR #1458 review notes
Document arkenv-internal exports, prefer typesafe wording, load optional
core engines via Jiti instead of createRequire, and lighten the Nuxt README
intro.
