+++ packages/nuxt/src/validate-schema.ts
@@ -1,28 +1,7 @@
-import { createRequire } from "node:module";
-import type { Dict, SchemaShape } from "@repo/types";
+import type { Dict } from "@repo/types";
 import type { BootGateEngine } from "./boot-gate";
 import { loadSchemaViaCapture } from "./boot-gate";
-
-const require = createRequire(import.meta.url);
-type CoreArkenv = (
-	schema: SchemaShape,
-	config?: { env?: Dict<string>; safe?: boolean },
-) => Record<string, unknown>;
-
-/**
- * Resolve the core validation function for the configured engine.
- *
- * @param engine ArkType (`@arkenv/core`) or Standard Schema (`@arkenv/standard`)
- * @returns The engine's `arkenv` function
- */
-function resolveCoreArkenv(engine: BootGateEngine): CoreArkenv {
-	if (engine === "standard") {
-		// eslint-disable-next-line @typescript-eslint/no-require-imports
-		return require("@arkenv/standard").arkenv as CoreArkenv;
-	}
-	// eslint-disable-next-line @typescript-eslint/no-require-imports
-	return require("@arkenv/core").arkenv as CoreArkenv;
-}
+import { resolveCoreArkenv } from "./resolve-core-arkenv";
 
 /**
  * Validate the project schema at build/dev time by loading schema via capture
